requirements.txt
# Development requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
kafka-python==2.0.2
neo4j==5.15.0
minio==7.2.0

# Machine Learning
lightgbm==4.1.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4
tensorflow==2.15.0
torch==2.1.2
networkx==3.2.1
node2vec==0.4.6

# Feature Engineering
geopy==2.4.1
python-dateutil==2.8.2
holidays==0.37

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.38.0

# Data Processing
pyarrow==14.0.2
requests==2.31.0
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.6.0
mypy==1.7.1

# Jupyter
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
typer==0.9.0
rich==13.7.0
tqdm==4.66.1
